Move functionality into Navbar
